微信读书答题辅助/requirements.txt
beautifulsoup4==4.8.0
urllib3==1.24.2
pillow==6.2.0
requests==2.22.0

